# docs/swagger.yaml
basePath: /gymtranning/v1/api
definitions:
...
  models.GymProfile:
    properties:
      address:
        type: string
      celPhone:
        type: string
      id:
        type: integer
      isWhatsApp:
        type: boolean
      name:
        type: string
      phone:
        type: string
    type: object
...
  models.UserProfile:
    properties:
      address:
        type: string
      celPhone:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
    type: object
  utils.Error:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
host: gymtrainning-api.herokuapp.com/
info:
  contact:
    email: marcelo.cheruti@gmail.com
    name: MCCAETANO
    url: http://www.swagger.io/support
  description: Show your trainning and manage your trainning
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Gym Trainning
  version: "1.0"
paths:
...
  /gym/profile:
    get:
      consumes:
      - application/json
      description: Get User
      operationId: UserProfileGet
      parameters:
      - description: Name of User
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.UserProfile'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Error'
        "404":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Error'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.Error'
      summary: Get User
      tags:
      - UserProfile
    post:
      consumes:
      - application/json
      description: Create Gym
      operationId: GymProfilePost
      parameters:
      - description: GymProfile Model
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.GymProfile'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GymProfile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Error'
        "404":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Error'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.Error'
      summary: Create Gym
      tags:
      - GymProfile
    put:
      consumes:
      - application/json
      description: Updade Gym
      operationId: GymProfilePut
      parameters:
      - description: GymProfile Model
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.GymProfile'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GymProfile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Error'
        "404":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Error'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.Error'
      summary: Updade Gym
      tags:
      - GymProfile
...
  /user/profile:
    post:
      consumes:
      - application/json
      description: Create User
      operationId: UserProfilePost
      parameters:
      - description: UserProfile Model
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UserProfile'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserProfile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Error'
        "404":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Error'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.Error'
      summary: Create User
      tags:
      - UserProfile
    put:
      consumes:
      - application/json
      description: Updade User
      operationId: UserProfilePut
      parameters:
      - description: UserProfile Model
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UserProfile'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserProfile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Error'
        "404":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Error'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.Error'
      summary: Updade User
      tags:
      - UserProfile
  /user/profile/{id}:
    delete:
      consumes:
      - application/json
      description: Delete User by id
      operationId: UserProfileDelete
      parameters:
      - description: Id of User
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserProfile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Error'
        "404":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Error'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.Error'
      summary: Delete User by id
      tags:
      - UserProfile
    get:
      consumes:
      - application/json
      description: Get User by id
      operationId: UserProfileGetById
      parameters:
      - description: Id of User
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserProfile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Error'
        "404":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Error'
        default:
          description: ""
          schema:
            $ref: '#/definitions/utils.Error'
      summary: Get User by id
      tags:
      - UserProfile
securityDefinitions:
  OAuth2Password:
    flow: password
    scopes:
      admin: ' Grants read and write access to administrative information'
      read: ' Grants read access'
      write: ' Grants write access'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
swagger: "2.0"